Fix checking for Sublime Text

+++ sublime/config.sh
@@ -2,14 +2,20 @@
 
 set -o errexit
 set -o nounset
-set -o xtrace
 
 SOURCE="/Applications/Sublime Text.app/Contents/SharedSupport/bin/subl"
 TARGET=/usr/local/bin/subl
 
+# Exit silently if not running on macOS
+# or if Sublime Text is not installed
+if [ $(uname) != Darwin ] || [ ! -e "$SOURCE" ]
+then
+    exit 0
+fi
+
+set -o xtrace
+
 which subl && exit 0
 
-test -e "$SOURCE"
 test ! -e "$TARGET"
-
 ln -s "$SOURCE" "$TARGET"
